requirements-test.txt
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
ruff>=0.3.0
mypy>=1.8.0
pandas-stubs>=2.1.0
httpx>=0.27.0
